delete: Comments accidentally committed and pushed
This is a useful function to create an admin user.

It should be removed with the credentials (for a db used for testing with fake data, moreover this created user has been removed, so it is not a security risk having it in the version control history)

A more preferred way to create a user with are wp-cli and sql queries. However, wp-cli method produced an error (with a completely useless message, no search results whatsoever in the whole internet) when trying to use it in this  Bedrock setup. TODO: Determine cause of the error.