fix: migrations run option updates
Run migrations option should be requested inside the for loop in every iteration, because we are updating the option array with new versions.
We need to check if those versions exist in the up-to-date array, not in the old one! Otherwise, we will have multiple identical versions in the run migrations option / array.